feat: consolidate CLI commands w/ enhanced tailor modes & unified help system
