+++ src/cli/commands/plan.py
@@ -29,9 +29,26 @@ from ..params import (
     ModelOpt,
     SectionsPathOpt,
 )
+from ...ui.help.help_data import command_help
 
 
 # * Generate edits w/ step-by-step planning & validation workflow
+@command_help(
+    name="plan",
+    description="Generate edits with step-by-step planning workflow (experimental)",
+    long_description=(
+        "Experimental command that uses a multi-step planning approach for "
+        "resume tailoring. Provides more detailed reasoning and step-by-step "
+        "edit generation for complex tailoring scenarios. Accepts the same "
+        "options as 'tailor' for risk and validation policies."
+    ),
+    examples=[
+        "loom plan job.txt resume.docx",
+        "loom plan job.txt resume.docx --sections-path sections.json",
+        "loom plan job.txt resume.docx --edits-json planned_edits.json",
+    ],
+    see_also=["tailor"],
+)
 @app.command(help="Generate edits with step-by-step planning workflow (experimental)")
 @handle_loom_error
 def plan(
